sidecar/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
chromadb>=0.4.0
boto3>=1.34.0
anthropic>=0.39.0
python-dotenv>=1.0.0
